backend/api_server/requirements.txt
Django==2.2.12
django-cors-headers==3.2.1
django-webpack-loader==0.7.0
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
pkg-resources==0.0.0
psycopg2==2.8.5
PyJWT==1.7.1
pytz==2020.1
sqlparse==0.3.1
